requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Deprecated==1.3.1
Flask==3.1.2
idna==3.11
img2pdf==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
packaging==25.0
pikepdf==10.0.0
pillow==12.0.0
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.5.0
validators==0.35.0
Werkzeug==3.1.3
wrapt==2.0.0
